requirements.txt
torch
pypdf 
langchain-community
langchain_community
faiss-cpu
langchain
transformers
sentence-transformers
pinecone-client